-- Implement support for toolchains into build system's base
This adds support for searching apropriate version of headers and libraries
via Find_*. There are 2 critical needs for building crtmpserver:
 OpenSSL
 DL
OpenSSL we must build for target platform manually and crtmpserver must find
it while configure project. Another, DL, must be provided by toolchain, and
cmake must find it inside toolchain too. If we do not use toolchain, variables
in Find_* is empty, so not any additianl path will be provided.